feat(frontmatter): add a lot of properties
Added type `FrontMatter` with the structure of data and metadata used in
each note.

The main feat is the boolean property `metadata.keep_in_sync` that
control if a note has to be updated or not.

Another useful metadata is `metadata.yet_to_be_edited` set equal to
`true` when the note is imported.

Right now the default for this settings is `false` and in a future
version it'll have an interface to be controlled.

Close #10